refactor: move window implementation from job to models package to avoid cyclic dependency
